Layer: Fix position using width/height when layer is opened.
This was done already for IE, but it turns out it is better for
all browsers. A once opened Layer should not change its postion
unless "position" is called. Without width/height set, changing
content inside the Layer resized the layer without adjusting the
anchor pointer and other side effects,

This patch also fixes a CSS problem in max-horizontal layouts inside
a vertical layout. This needs to be width: 100% which it wasnt.

Removed code from Tab.coffee which was unreachable for CUI.__ng__
(which we have always set to true).