Layer: Added ".cui-layer-sidebar--active" class marker.
This marker allows users of CUI to tell the Layer system that the viewport
for positioning has an active left sidebar. CUI measures the DIV found
with that class and adjusts all positioning to it.

Pane.fillScreen also acknowledges such sidebar, so make sure you
don't have an "end sidebar" button visible in the sidebar, as the Pane
fill screen would not notice the closed sidebar.